Sun and solar objects (#47)
* Refactor: Rename Planets to SolarObjects

The name "Planets" is not entirely accurate, as it includes the Moon.
This commit renames the class to "SolarObjects" to better reflect its
contents. The notification template and tests are updated as well.

* Normalize planet size for plotting

* Fix time bounds for weather and observation data

* Corrects the bounds for the `get_critical_data` method in `Weather`
and the time filtering in `Observation`. * Adds debugging logs for the
time ranges of downloaded data. * Updates tests to reflect the time
bound changes.

* Refactor observation window calculations

- Renamed `start_time_for_observation_window` and
`stop_time_for_observation_window` to `start` and `stop`. - Updated the
`_normalize_dates` method call. - Removed the `hours` parameter from the
`_filter_data` method in `Weather` class and removed its usage. -
Updated tests to reflect these changes.

* Fix incorrect expected processed hours

The previous commit calculated the expected processed hours incorrectly
in `TestObservationWeatherAnalysis.test_within_time_limit`. The test now
correctly expects 8 hours to be processed.

* Fix KeyError in weather data parsing

Handle the case where the "hourly" or "data" keys are missing from the
JSON response by catching a KeyError and returning an empty DataFrame
with the expected columns. Also, log the error for debugging purposes.

* Fix visibility plot data filtering

Removes the `hours` parameter from the `_filter_data` call. This was
causing an issue where no data was being returned for the visibility
plot.

* Fix: Pass timezone information from get_hourly_weather

The timezone is now correctly returned from get_hourly_weather_analysis.
This change also updates tests to verify the timezone is being returned
correctly.

* Fix: Include time_limit hour in hourly weather analysis

The time_limit was previously exclusive, which meant the hour at the
time_limit was excluded from analysis. Now it is inclusive, ensuring
that the data for the time_limit hour is included.

Also, refactor tests to test this inclusive behavior.